refactor: merge후 홈페이지와 채널페이지 UI 연결하기
